+++ LibHac/Save/DuplexBitmap.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.IO;
 
-namespace LibHac.Savefile
+namespace LibHac.Save
 {
     public class DuplexBitmap
     {
