+++ LibHac/Save/Header.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 
-namespace LibHac.Savefile
+namespace LibHac.Save
 {
     public class Header
     {
